# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BidRush Auction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/auctions:
    get:
      operationId: getMyAuctions
      tags: [users]
      summary: Get current user's auctions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User auctions
          content:
            application/json:
              schema:
                type: object
                properties:
                  auctions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Auction"
                required:
                  - auctions
  /users/me/bids:
    get:
      operationId: getMyBids
      tags: [users]
      summary: Get current user's bids
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User bids
          content:
            application/json:
              schema:
                type: object
                properties:
                  bids:
                    type: array
                    items:
                      $ref: "#/components/schemas/BidWithAuction"
                required:
                  - bids
  /users/me/won:
    get:
      operationId: getMyWonAuctions
      tags: [users]
      summary: Get auctions user has won
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Won auctions
          content:
            application/json:
              schema:
                type: object
                properties:
                  auctions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Auction"
                required:
                  - auctions
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                required:
                  - notifications
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
          nullable: true
        rating:
          type: number
        totalSales:
          type: integer
        totalPurchases:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - rating
        - totalSales
        - totalPurchases
        - createdAt
    Auction:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        startingPrice:
          type: number
        currentPrice:
          type: number
        minimumIncrement:
          type: number
        status:
          type: string
          enum: [live, scheduled, ended]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        sellerId:
          type: string
        sellerName:
          type: string
        sellerAvatar:
          type: string
          nullable: true
        bidCount:
          type: integer
        winnerId:
          type: string
          nullable: true
        winnerName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - images
        - category
        - startingPrice
        - currentPrice
        - minimumIncrement
        - status
        - startTime
        - endTime
        - sellerId
        - sellerName
        - bidCount
        - createdAt
...
    Bid:
      type: object
      properties:
        id:
          type: string
        auctionId:
          type: string
        bidderId:
          type: string
        bidderName:
          type: string
        bidderAvatar:
          type: string
          nullable: true
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - auctionId
        - bidderId
        - bidderName
        - amount
        - createdAt
    BidWithAuction:
      allOf:
        - $ref: "#/components/schemas/Bid"
        - type: object
          properties:
            auction:
              $ref: "#/components/schemas/Auction"
          required:
            - auction
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [outbid, auction_ending, auction_won, auction_lost, new_bid]
        title:
          type: string
        message:
          type: string
        auctionId:
          type: string
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - message
        - read
        - createdAt